fix database

+++ backend/database/migrations/2026_07_02_041118_add_ticket_print_fields_to_fb_products_table.php
@@ -6,23 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::table('fb_products', function (Blueprint $table) {
-            $table->integer('is_in_stock')->default(1)->after('track_stock');
+            if (!Schema::hasColumn('fb_products', 'is_in_stock')) {
+                $table->integer('is_in_stock')->default(1)->after('track_stock');
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('fb_products', function (Blueprint $table) {
-            $table->dropColumn('is_in_stock');
+            if (Schema::hasColumn('fb_products', 'is_in_stock')) {
+                $table->dropColumn('is_in_stock');
+            }
         });
     }
 };
